requirements.txt
streamlit~=1.37.1
numpy~=1.26.4
tensorflow~=2.18.0
pillow~=10.4.0